+++ game/pieces.py
@@ -1,4 +1,4 @@
-class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bishop, Knight, Position = (row , column)
+class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bishop, Knight, Position = (row , column) ♜
     def __init__(self, piece, color, initial_position):
         self.__name__ = piece
         self.__color__ = color
@@ -7,20 +7,20 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
     
     def set_images(self):
         images_white = {
-                        'Queen': '♛',
-                        'King': '♚',
-                        'Pawn' : '♗',
-                        'Knight': '♞',
-                        'Rook': '♜',
-                        'Bishop' : '♝'
+                        'Queen': 'q',
+                        'King': 'k',
+                        'Pawn' : 'p',
+                        'Knight': 'n',
+                        'Rook': 'r',
+                        'Bishop' : 'b'
                         }
         images_black = {
-                        'Queen': '♕',
-                        'King': '♔',
-                        'Pawn' : '♙',
-                        'Knight': '♘',
-                        'Rook': '♖',
-                        'Bishop' : '♗'
+                        'Queen': 'Q',
+                        'King': 'K',
+                        'Pawn' : 'P',
+                        'Knight': 'N',
+                        'Rook': 'R',
+                        'Bishop' : 'B'
                         }
         if self.__color__ == 'w':
             self.__image__ = images_white[self.__name__]